scraper/requirements.txt
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0